pyproject.toml
[project]
name = "axonpush"
version = "0.0.10"
description = "Python SDK for AxonPush — real-time event infrastructure for AI agent systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "AxonPush" }]
keywords = [
    "axonpush",
    "ai-agents",
    "real-time",
    "events",
    "tracing",
    "observability",
    "webhooks",
    "mqtt",
    "iot",
    "langchain",
    "langgraph",
    "deepagents",
    "openai",
    "anthropic",
    "crewai",
    "multi-agent",
    "distributed-tracing",
    "sdk",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
    "Framework :: Pydantic :: 2",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "httpx>=0.25.0,<1.0",
    "pydantic>=2.0,<3.0",
    "paho-mqtt>=2.0,<3.0",
    "aiomqtt>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
]

[project.urls]
Homepage = "https://github.com/axonpush/python-sdk"
Repository = "https://github.com/axonpush/python-sdk"
Issues = "https://github.com/axonpush/python-sdk/issues"
Changelog = "https://github.com/axonpush/python-sdk/releases"

[project.optional-dependencies]
langchain = ["langchain-core>=0.1.0"]
openai-agents = ["openai-agents>=0.1.0"]
anthropic = ["anthropic>=0.30.0"]
crewai = ["crewai>=0.50.0; python_version>='3.11'"]
deepagents = ["deepagents>=0.1.0; python_version>='3.11'"]
loguru = ["loguru>=0.7.0"]
structlog = ["structlog>=24.0.0"]
rq = ["rq>=2.0"]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
all = [
    "axonpush[langchain]",
    "axonpush[openai-agents]",
    "axonpush[anthropic]",
    "axonpush[crewai]",
    "axonpush[deepagents]",
    "axonpush[rq]",
    "axonpush[loguru]",
    "axonpush[structlog]",
    "axonpush[otel]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axonpush"]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["examples"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "agents",
    "agents.*",
    "anthropic",
    "anthropic.*",
    "crewai",
    "crewai.*",
    "deepagents",
    "deepagents.*",
    "langchain_core",
    "langchain_core.*",
    "loguru",
    "loguru.*",
    "opentelemetry",
    "opentelemetry.*",
    "redis",
    "redis.*",
    "rq",
    "rq.*",
    "sentry_sdk",
    "sentry_sdk.*",
    "structlog",
    "structlog.*",
]
ignore_missing_imports = true

# Integrations subclass third-party callback / hook classes whose
# stubs we don't control. Allow ``class Foo(ThirdPartyAny): ...`` here.
[[tool.mypy.overrides]]
module = [
    "axonpush.integrations.anthropic",
    "axonpush.integrations.crewai",
    "axonpush.integrations.deepagents",
    "axonpush.integrations.langchain",
    "axonpush.integrations.openai_agents",
    "axonpush.integrations.otel",
]
disallow_subclassing_any = false

# Resource modules dispatch through the facade's ``_invoke`` chokepoint,
# which is typed ``Any`` in the Protocol (op arg is a duck-typed module).
# Each method signature declares the precise return; allow the bare-Any
# return inside the function body so we don't have to ``cast`` 60+ sites.
[[tool.mypy.overrides]]
module = ["axonpush.resources.*"]
disable_error_code = ["no-any-return"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not e2e'"
markers = [
    "e2e: end-to-end tests requiring a live easy-push backend (opt in with `pytest -m e2e`)",
]

[dependency-groups]
dev = [
    "openapi-python-client>=0.21,<1.0",
    "types-python-dateutil>=2.9.0.20260408",
]
